Build an abstract class for node_resolve and node_paths to make distinction more clear
