+++ contrib/node/src/python/pants/contrib/node/tasks/node_paths.py
@@ -4,29 +4,29 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from builtins import object
+from pants.util.meta import AbstractClass
 
 
 # TODO(John Sirois): UnionProducts? That seems broken though for ranged version constraints,
 # which npm has and are widely used in the community.  For now stay dumb simple (and slow) and
 # resolve each node_module individually.
-class NodePaths(object):
-  """Maps NpmPackage targets to their resolved NODE_PATH chroot."""
+class NodePathsBase(AbstractClass):
+  """Maps NpmPackage targets to their resolved NODE_PATH."""
 
   def __init__(self):
     self._paths_by_target = {}
 
   def resolved(self, target, node_path):
-    """Identifies the given target as resolved to the given chroot path.
+    """Identifies the given target as resolved to the given path.
 
     :param target: The target that was resolved to the `node_path` chroot.
     :type target: :class:`pants.contrib.node.targets.npm_package.NpmPackage`
-    :param string node_path: The chroot path the given `target` was resolved to.
+    :param string node_path: The path the given `target` was resolved to.
     """
     self._paths_by_target[target] = node_path
 
   def node_path(self, target):
-    """Returns the path of the resolved chroot for the given NpmPackage.
+    """Returns the path of the resolved root for the given NpmPackage.
 
     Returns `None` if the target has not been resolved to a chroot.
 
@@ -41,3 +41,23 @@ class NodePaths(object):
     :rtype list string
     """
     return list(self._paths_by_target.values())
+
+
+class NodePaths(NodePathsBase):
+  """Maps NpmPackage targets to their resolved NODE_PATH chroot.
+
+  NodePaths is exposed as a product type for resolve.node.
+  The paths are the resolved chroot after copying sources to the synthetic workspace.
+  """
+
+
+class NodePathsLocal(NodePathsBase):
+  """Maps Node package targets to their local NODE_PATH chroot.
+
+  NodePathsLocal is exposed as a product type for node-resolve-local.
+  The paths are the real source chroot for the target.
+
+  This product is intended to be used with the node-install goal which will
+  install local and 3rd party dependencies into the source directory rather
+  than the pants working directory.
+  """
