+++ contrib/node/src/python/pants/contrib/node/tasks/node_resolve.py
@@ -4,14 +4,18 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from pants.base.workunit import WorkUnitLabel
+import os
 
-from pants.contrib.node.tasks.node_paths import NodePaths
+from pants.base.build_environment import get_buildroot
+from pants.base.workunit import WorkUnitLabel
+from pants.util.meta import AbstractClass
+
+from pants.contrib.node.tasks.node_paths import NodePaths, NodePathsLocal
 from pants.contrib.node.tasks.node_task import NodeTask
 
 
-class NodeResolve(NodeTask):
-  """Resolves node_package targets to isolated chroots using different registered resolvers."""
+class NodeResolveBase(NodeTask, AbstractClass):
+  """Resolves node_package targets to their node paths using different registered resolvers."""
 
   _resolver_by_type = dict()
 
@@ -22,7 +26,7 @@ class NodeResolve(NodeTask):
   @classmethod
   def prepare(cls, options, round_manager):
     """Allow each resolver to declare additional product requirements."""
-    super(NodeResolve, cls).prepare(options, round_manager)
+    super(NodeResolveBase, cls).prepare(options, round_manager)
     for resolver in cls._resolver_by_type.values():
       resolver.prepare(options, round_manager)
 
@@ -67,6 +71,13 @@ class NodeResolve(NodeTask):
     return self.is_node_package(target) and self._resolver_for_target(target) != None
 
   def execute(self):
+    pass
+
+
+class NodeResolve(NodeResolveBase):
+  """Resolves node_package targets to isolated chroots using different registered resolvers."""
+
+  def execute(self):
     targets = self.context.targets(predicate=self._can_resolve_target)
     if not targets:
       return
@@ -86,3 +97,53 @@ class NodeResolve(NodeTask):
             resolver_for_target_type = self._resolver_for_target(target).global_instance()
             resolver_for_target_type.resolve_target(self, target, vt.results_dir, node_paths)
           node_paths.resolved(target, vt.results_dir)
+
+
+class NodeResolveLocal(NodeResolveBase):
+  """Resolves node_package targets to source chroots using different registered resolvers.
+
+  This task sets up the development environment
+  """
+
+  @classmethod
+  def product_types(cls):
+    # NodePathsLocal contain Node Paths that are local to the source target root.
+    return [NodePathsLocal]
+
+  @property
+  def cache_target_dirs(self):
+    # Do not create a results_dir for this task
+    return True
+
+  def artifact_cache_reads_enabled(self):
+    # Artifact caching is not necessary for local installation.
+    # Just depend on the local cache provided by the package manager.
+    return False
+
+  def execute(self):
+    targets = self.context.targets(predicate=self._can_resolve_target)
+    if not targets:
+      return
+
+    node_paths = self.context.products.get_data(NodePathsLocal, init_func=NodePathsLocal)
+    # Invalidate all targets for this task, and use invalidated() check
+    # to build topologically sorted target graph. This is probably not the best way
+    # to do this, but it works.
+    self.invalidate()
+    with self.invalidated(targets,
+                          # This is necessary to ensure that transitive dependencies are installed first
+                          topological_order=True,
+                          invalidate_dependents=True,
+                          silent=True) as invalidation_check:
+      with self.context.new_workunit(name='node-install', labels=[WorkUnitLabel.MULTITOOL]):
+        for vt in invalidation_check.all_vts:
+          target = vt.target
+          if not vt.valid:
+            resolver_for_target_type = self._resolver_for_target(target).global_instance()
+            results_dir = os.path.join(get_buildroot(), target.address.spec_path)
+            resolver_for_target_type.resolve_target(self, target, results_dir, node_paths,
+                                                    resolve_locally=True,
+                                                    force=True,
+                                                    install_optional=True,
+                                                    frozen_lockfile=False)
+          node_paths.resolved(target, results_dir)
